+++ libs/translator/reader/read_shader.cpp
@@ -212,10 +212,13 @@ AtNode* UsdArnoldReadShader::Read(const UsdPrim &prim, UsdArnoldReaderContext &c
 
     const std::string& nodeName = prim.GetPath().GetString();
     // Ensure we don't re-export a shader that was already exported 
-    // (which can happen when a shader is connected multiple times)
-    AtNode* node = context.GetReader()->LookupNode(nodeName.c_str());
-    if (node) {
-        return node;
+    // (which can happen when a shader is connected multiple times).
+    // However, if we're doing an interactive update, we cannot skip this.
+    if (!context.GetReader()->IsUpdating()) {
+        AtNode* node = context.GetReader()->LookupNode(nodeName.c_str());
+        if (node) {
+            return node;
+        }
     }
 
     UsdShadeShader shader(prim);